backend/requirements.txt
fastapi[standard]==0.116.1
SQLAlchemy[asyncio]==2.0.43
asyncpg==0.30.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.2.2
pytest-asyncio==0.23.8
asgi-lifespan==2.1.0
pytest-cov==5.0.0
certifi==2024.8.30
numpy==2.1.2
python-dateutil==2.9.0.post0
scipy==1.14.1